// src/components/Button/Button.types.ts
import { MouseEventHandler } from "react";
import { ButtonSizes } from "../../types/Sizes";
import { ButtonVariants } from "../../types/Variants";
export interface ButtonProps {
  text?: string;
  variant?: ButtonVariants;
  disabled?: boolean;
  loading?: boolean;
  size?: ButtonSizes;
  onClick?: MouseEventHandler<HTMLButtonElement>;
}